# docs/api.yml
openapi: 3.0.3
info:
  title: Restaurant API
  description: STD23042
  version: '1.0.0'
servers:
  - url: http://localhost:8080
    description: Local server
...
paths:
...
  /dishes/{id}/ingredients:
    put:
      tags:
        - Dishes
      summary: Add or update ingredients for a dish
      operationId: updateDishes
      parameters:
        - name: dishId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: array
              items:
                $ref: '#/components/schemas/DishIngredient'
      responses:
        '200':
          description: Updated dish
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Dish'
  /orders/{reference}:
    get:
      tags:
        - Orders
      summary: Get order by reference
      operationId: getOrders
      parameters:
        - name: reference
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: get Order
          content:
            application/json:
              schema:
                  $ref: '#/components/schemas/Order'
  /orders/{reference}/dishes:
    put:
      tags:
        - Orders
      summary: Add or update dishes for an order
      operationId: updateDishOrders
      parameters:
        - name: reference
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: array
              items:
                $ref: '#/components/schemas/CreateDishOrder'
      responses:
        '200':
          description: Updated orders
          content:
            application/json:
              schema:
                  $ref: '#/components/schemas/Order'
  /orders/{reference}/dishes/{dishId}:
    put:
      tags:
        - Dish Status
      summary: Update status for a dish in order
      operationId: updateDishStatus
      parameters:
        - name: reference
          in: path
          required: true
          schema:
            type: string
        - name: dishId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: array
              items:
                $ref: '#/components/schemas/DishOrder'
      responses:
        '200':
          description: Updated orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/DishOrder'
components:
  schemas:
    Ingredient:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        updateDatetime:
          type: string
          format: date-time
        unitPrice:
          type: number
          format: double
        unit:
          type: string
        priceHistories:
          type: array
          items:
            $ref: '#/components/schemas/PriceHistory'
        ingredientStockMovements:
          type: array
          items:
            $ref: '#/components/schemas/IngredientStockMovement'
    PriceHistory:
      type: object
      properties:
        id:
          type: string
        idIngredient:
          type: string
        priceDatetime:
          type: string
          format: date-time
        unitPrice:
          type: number
          format: double
    IngredientStockMovement:
      type: object
      properties:
        id:
          type: string
        idIngredient:
          type: string
        quantity:
          type: number
          format: float
        movementDatetime:
          type: string
          format: date-time
        movementType:
          type: string
        unit:
          type: string
    Dish:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        availableQuantity:
          type: integer
        price:
          type: number
          format: big decimal
        ingredient:
          type: array
          items:
            $ref: '#/components/schemas/Ingredient'
    DishIngredient:
      type: object
      properties:
        requiredQuantity:
          type: number
          format: float
        unit:
          $ref: '#/components/schemas/Unit'
        idIngredient:
          type: string
    Unit:
      type: string
      enum:
        - G
        - L
        - U
    StatusHistory:
      type: string
      enum:
        - CREATED
        - CONFIRMED
        - IN_PREPARATION
        - COMPLETED
        - SERVED
    DishOrderStatus:
      type: object
      properties:
        id:
          type: string
        dishOrderId:
          type: string
        status:
          $ref: '#/components/schemas/StatusHistory'
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    DishOrder:
      type: object
      properties:
        id:
          type: string
        orderId:
          type: string
        dish:
          $ref: '#/components/schemas/Dish'
        quantity:
          type: integer
        dishOrderStatus:
          $ref: '#/components/schemas/DishOrderStatus'
    OrderStatus:
      type: object
      properties:
        id:
          type: string
        orderId:
          type: string
        status:
          $ref: '#/components/schemas/StatusHistory'
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    Order:
      type: object
      properties:
        id:
          type: string
        reference:
          type: string
        createAt:
          type: string
          format: date-time
        updated:
          $ref: '#/components/schemas/OrderStatus'
        orderStatus:
          type: number
          format: big decimal
        dishOrder:
          type: array
          items:
            $ref: '#/components/schemas/DishOrder'
    CreateDishOrder:
      type: object
      properties:
        id:
          type: string
        dishId:
          type: string
        quantity:
          type: integer
        status:
          $ref: '#/components/schemas/StatusHistory'